Laptop CRUD Operation with react

+++ laptop/src/main/java/com/xworkz/laptop/service/LaptopServiceImpl.java
@@ -3,13 +3,15 @@ package com.xworkz.laptop.service;
 import com.xworkz.laptop.dto.LaptopDto;
 import com.xworkz.laptop.entity.LaptopEntity;
 import com.xworkz.laptop.repository.LaptopReository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-
+import java.util.Optional;
+@Slf4j
 @Service
 public class LaptopServiceImpl implements  LaptopService{
     private final LaptopReository laptopReository;
@@ -35,8 +37,14 @@ public class LaptopServiceImpl implements  LaptopService{
 
     @Override
     public boolean deleteLaptop(int id) {
-        laptopReository.deleteById(id);
-        return true;
+       Optional<LaptopEntity> laptopEntity =laptopReository.findById(id);
+       if(laptopEntity.isPresent()){
+           laptopReository.deleteById(id);
+           return true;
+       }else{
+           return false;
+       }
+
     }
 
     @Override
@@ -52,5 +60,35 @@ public class LaptopServiceImpl implements  LaptopService{
             return laptopDtos;
         }
         return Collections.emptyList();
+    }
+
+    @Override
+    public LaptopDto findById(int id) {
+        Optional<LaptopEntity> laptopEntity= laptopReository.findById(id);
+        if(laptopEntity!=null &&laptopEntity.isPresent()){
+            com.xworkz.laptop.dto.LaptopDto laptopDto=new LaptopDto();
+            BeanUtils.copyProperties(laptopEntity.get(),laptopDto);
+            log.info(laptopDto.toString());
+            return laptopDto;
+        }else {
+            return null;
+        }
+    }
+
+    @Override
+    public boolean updateLaptop(LaptopDto laptopDto) {
+      Optional<LaptopEntity> laptopEntity=  laptopReository.findById(laptopDto.getId());
+        if (laptopEntity.isPresent()) {
+            LaptopEntity laptop = laptopEntity.get();
+            laptop.setLaptopName(laptopDto.getLaptopName());
+            laptop.setLaptopBrand(laptopDto.getLaptopBrand());
+            laptop.setLaptopPrice(laptopDto.getLaptopPrice());
+            laptop.setOwnerName(laptopDto.getOwnerName());
+
+            laptopReository.save(laptop);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
